+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Examples.hs
@@ -5,15 +5,16 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
+-- | The example transactions in this module are not valid transactions. We
+-- don't care, we are only interested in serialisation, not validation.
 module Test.Cardano.Ledger.Babbage.Examples (
   ledgerExamples,
-  mkBabbageBasedExampleTx,
-  exampleBabbageBasedTxBody,
   exampleBabbageNewEpochState,
+  exampleBabbageBasedTx,
+  exampleBabbageBasedTopTx,
 ) where
 
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
-import Cardano.Ledger.Alonzo.Scripts (AlonzoPlutusPurpose (..))
 import Cardano.Ledger.Babbage (ApplyTxError (BabbageApplyTxError), BabbageEra)
 import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.BaseTypes (StrictMaybe (..))
@@ -34,14 +35,16 @@ import Cardano.Ledger.Shelley.Rules (
  )
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
+import qualified Data.MapExtras as Map
 import qualified Data.Sequence.Strict as StrictSeq
+import Data.Typeable (Typeable)
 import Lens.Micro
 import Test.Cardano.Ledger.Alonzo.Arbitrary (alwaysSucceeds)
 import Test.Cardano.Ledger.Alonzo.Examples (
-  exampleAlonzoBasedShelleyTxBody,
-  exampleAlonzoBasedTxBody,
+  addAlonzoToConwayExampleReqSigners,
+  exampleAlonzoBasedTopTx,
+  exampleAlonzoBasedTx,
   exampleDatum,
-  mkAlonzoBasedExampleTx,
   mkAlonzoBasedLedgerExamples,
  )
 import Test.Cardano.Ledger.Babbage.Era
@@ -50,8 +53,13 @@ import Test.Cardano.Ledger.Mary.Examples (exampleMultiAssetValue)
 import Test.Cardano.Ledger.Plutus (alwaysSucceedsPlutus)
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples (..),
+  addShelleyBasedTopTxExampleFee,
+  addShelleyToBabbageExampleProposedPUpdates,
+  addShelleyToBabbageTxCerts,
+  addShelleyToConwayTxCerts,
   exampleNewEpochState,
   examplePayKey,
+  exampleShelleyScript,
   exampleStakeKey,
   exampleTxIns,
   mkKeyHash,
@@ -68,11 +76,17 @@ ledgerExamples =
                 DelegateeNotRegisteredDELEG (mkKeyHash 1)
     )
     exampleBabbageNewEpochState
-    ( mkBabbageBasedExampleTx
-        exampleBabbageBasedShelleyTxBody
-        (AlonzoSpending $ AsIx 0)
-    )
+    exampleBabbageTx
     NoGenesis
+  where
+    exampleBabbageTx :: Tx TopTx BabbageEra
+    exampleBabbageTx =
+      exampleBabbageBasedTopTx
+        & addShelleyBasedTopTxExampleFee
+        & addShelleyToBabbageExampleProposedPUpdates
+        & addShelleyToBabbageTxCerts
+        & addShelleyToConwayTxCerts
+        & addAlonzoToConwayExampleReqSigners
 
 exampleBabbageNewEpochState ::
   ( BabbageEraTest era
@@ -85,83 +99,91 @@ exampleBabbageNewEpochState =
     emptyPParams
     (emptyPParams & ppCoinsPerUTxOByteL .~ CoinPerByte (CompactCoin 1))
 
-mkBabbageBasedExampleTx ::
+exampleBabbageBasedTopTx ::
   forall era.
   ( AlonzoEraTx era
+  , BabbageEraTxBody era
   , AlonzoEraTxAuxData era
-  , EraPlutusTxInfo 'PlutusV1 era
-  , EraPlutusTxInfo 'PlutusV2 era
+  , Value era ~ MaryValue
+  , EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
   ) =>
-  TxBody TopTx era ->
-  PlutusPurpose AsIx era ->
   Tx TopTx era
-mkBabbageBasedExampleTx txBody scriptPurpose =
-  mkAlonzoBasedExampleTx txBody scriptPurpose
+exampleBabbageBasedTopTx =
+  exampleAlonzoBasedTopTx
+    & addBabbageBasedTxFeatures
+    & addBabbageBasedTopTxFeatures
+
+exampleBabbageBasedTx ::
+  forall era l.
+  ( AlonzoEraTx era
+  , BabbageEraTxBody era
+  , AlonzoEraTxAuxData era
+  , Value era ~ MaryValue
+  , EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
+  , Typeable l
+  ) =>
+  Tx l era
+exampleBabbageBasedTx =
+  exampleAlonzoBasedTx
+    & addBabbageBasedTxFeatures
+
+addBabbageBasedTopTxFeatures ::
+  forall era.
+  ( AlonzoEraTx era
+  , BabbageEraTxBody era
+  , Value era ~ MaryValue
+  ) =>
+  Tx TopTx era ->
+  Tx TopTx era
+addBabbageBasedTopTxFeatures tx =
+  tx
+    & bodyTxL . collateralReturnTxBodyL .~ SJust exampleCollateralOutput
+    & bodyTxL . totalCollateralTxBodyL .~ SJust (Coin 8675309)
+
+addBabbageBasedTxFeatures ::
+  forall era l.
+  ( AlonzoEraTx era
+  , BabbageEraTxBody era
+  , AlonzoEraTxAuxData era
+  , Value era ~ MaryValue
+  , EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
+  ) =>
+  Tx l era ->
+  Tx l era
+addBabbageBasedTxFeatures tx =
+  tx
     & witsTxL
       <>~ ( mkBasicTxWits
-              & scriptTxWitsL
-                .~ Map.singleton
-                  (hashScript @era $ alwaysSucceeds @'PlutusV2 3)
-                  (alwaysSucceeds @'PlutusV2 3)
+              & scriptTxWitsL <>~ Map.fromElems hashScript [alwaysSucceeds @'PlutusV2 3]
           )
-    & auxDataTxL
-      %~ fmap
-        ( \auxData ->
-            auxData
-              & plutusScriptsTxAuxDataL
-                <>~ Map.singleton PlutusV2 (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV2 3)))
-        )
-
-exampleBabbageBasedShelleyTxBody ::
-  forall era.
-  ( BabbageEraTxBody era
-  , ShelleyEraTxBody era
-  , EraPlutusTxInfo PlutusV1 era
-  , EraPlutusTxInfo PlutusV2 era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era
-exampleBabbageBasedShelleyTxBody =
-  mkBabbageBasedExampleTxBody exampleAlonzoBasedShelleyTxBody
-
-exampleBabbageBasedTxBody ::
-  forall era.
-  ( BabbageEraTxBody era
-  , EraPlutusTxInfo PlutusV1 era
-  , EraPlutusTxInfo PlutusV2 era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era
-exampleBabbageBasedTxBody =
-  mkBabbageBasedExampleTxBody exampleAlonzoBasedTxBody
-
-mkBabbageBasedExampleTxBody ::
-  forall era.
-  ( BabbageEraTxBody era
-  , Value era ~ MaryValue
-  , EraPlutusTxInfo PlutusV1 era
-  , EraPlutusTxInfo PlutusV2 era
-  ) =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkBabbageBasedExampleTxBody txBody =
-  txBody
-    & referenceInputsTxBodyL .~ exampleTxIns
-    & outputsTxBodyL
+    & modifyTxAuxData
+      ( plutusScriptsTxAuxDataL
+          %~ Map.insertWith
+            (<>)
+            PlutusV2
+            (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV2 3)))
+      )
+    & bodyTxL . referenceInputsTxBodyL <>~ exampleTxIns
+    & bodyTxL . outputsTxBodyL
       <>~ StrictSeq.fromList
         [ mkBasicTxOut
             (mkAddr examplePayKey exampleStakeKey)
-            (exampleMultiAssetValue 2)
-            & datumTxOutL .~ Datum (dataToBinaryData exampleDatum) -- inline datum
+            (exampleMultiAssetValue 1)
+            & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
             & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV1 3)
         , mkBasicTxOut
             (mkAddr examplePayKey exampleStakeKey)
             (exampleMultiAssetValue 2)
-            & datumTxOutL .~ Datum (dataToBinaryData exampleDatum) -- inline datum
+            & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
             & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV2 3)
+        , mkBasicTxOut
+            (mkAddr examplePayKey exampleStakeKey)
+            (exampleMultiAssetValue 3)
+            & referenceScriptTxOutL .~ SJust (fromNativeScript exampleShelleyScript)
         ]
-    & collateralReturnTxBodyL .~ SJust exampleCollateralOutput
-    & totalCollateralTxBodyL .~ SJust (Coin 8675309)
 
 exampleCollateralOutput ::
   ( BabbageEraTxOut era