+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Examples.hs
@@ -1,39 +1,42 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 
+-- | The example transactions in this module are not valid transactions. We
+-- don't care, we are only interested in serialisation, not validation.
 module Test.Cardano.Ledger.Dijkstra.Examples (
   ledgerExamples,
-  mkDijkstraBasedExampleTx,
-  mkDijkstraBasedExampleTxBody,
+  exampleDijkstraBasedTopTx,
+  exampleDijkstraBasedSubTx,
 ) where
 
 import Cardano.Ledger.Address (DirectDeposits (..))
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
+import Cardano.Ledger.Alonzo.Scripts (ExUnits (..))
+import Cardano.Ledger.Alonzo.TxWits (Redeemers (..))
 import Cardano.Ledger.BaseTypes (
-  EpochNo (..),
   Exclusive (..),
   Inclusive (..),
   Network (..),
-  SlotNo (..),
   StrictMaybe (..),
  )
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Rules (ConwayDelegPredFailure (..))
-import Cardano.Ledger.Conway.TxCert (ConwayGovCert (..), Delegatee (..))
 import Cardano.Ledger.Credential (Credential (..))
-import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Dijkstra (ApplyTxError (..), DijkstraEra)
 import qualified Cardano.Ledger.Dijkstra.Rules as Dijkstra
 import Cardano.Ledger.Dijkstra.Scripts (
   AccountBalanceInterval (..),
   AccountBalanceIntervals (..),
-  DijkstraPlutusPurpose (..),
+  DijkstraEraScript,
+  pattern GuardingPurpose,
  )
 import Cardano.Ledger.Dijkstra.TxBody (
   DijkstraEraTxBody,
@@ -43,45 +46,37 @@ import Cardano.Ledger.Dijkstra.TxBody (
   requiredTopLevelGuardsL,
   subTransactionsTxBodyL,
  )
-import Cardano.Ledger.Dijkstra.TxCert
 import Cardano.Ledger.Mary.Value (MaryValue (..))
 import Cardano.Ledger.Plutus.Data (
+  Data (..),
   Datum (..),
   dataToBinaryData,
  )
 import Cardano.Ledger.Plutus.Language (Language (..), plutusBinary)
-import Cardano.Ledger.State (StakePoolParams (sppId))
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
+import qualified Data.MapExtras as Map
 import qualified Data.OMap.Strict as OMap
 import qualified Data.OSet.Strict as OSet
 import qualified Data.Sequence.Strict as StrictSeq
 import Lens.Micro ((%~), (&), (.~), (<>~))
+import qualified PlutusLedgerApi.Common as P
 import Test.Cardano.Ledger.Alonzo.Arbitrary (alwaysSucceeds)
 import Test.Cardano.Ledger.Alonzo.Examples (
   exampleDatum,
   mkAlonzoBasedLedgerExamples,
  )
 import Test.Cardano.Ledger.Babbage.Examples (exampleBabbageNewEpochState)
-import Test.Cardano.Ledger.Conway.Examples (
-  exampleAnchor,
-  exampleConwayBasedTxBody,
-  exampleProposalProcedure,
-  exampleVotingProcedures,
-  mkConwayBasedExampleTx,
- )
+import Test.Cardano.Ledger.Conway.Examples (exampleConwayBasedTopTx, exampleConwayBasedTx)
 import Test.Cardano.Ledger.Core.KeyPair (mkAddr)
-import Test.Cardano.Ledger.Core.Utils (mkDummySafeHash)
 import Test.Cardano.Ledger.Dijkstra.ImpTest (exampleDijkstraGenesis)
-import Test.Cardano.Ledger.Mary.Examples (exampleMultiAsset, exampleMultiAssetValue)
+import Test.Cardano.Ledger.Mary.Examples (exampleMultiAssetValue)
 import Test.Cardano.Ledger.Plutus (alwaysSucceedsPlutus)
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples (..),
+  addShelleyBasedTopTxExampleFee,
   examplePayKey,
   exampleStakeKey,
-  exampleStakePoolParams,
-  exampleTxIns,
-  exampleWithdrawals,
   mkKeyHash,
   mkScriptHash,
  )
@@ -96,59 +91,110 @@ ledgerExamples =
               DelegateeStakePoolNotRegisteredDELEG (mkKeyHash 1)
     )
     exampleBabbageNewEpochState
-    exampleTxDijkstra
+    exampleDijkstraTx
     exampleDijkstraGenesis
+  where
+    exampleDijkstraTx :: Tx TopTx DijkstraEra
+    exampleDijkstraTx =
+      exampleDijkstraBasedTopTx
+        & addShelleyBasedTopTxExampleFee
 
-exampleTxDijkstra :: Tx TopTx DijkstraEra
-exampleTxDijkstra =
-  mkDijkstraBasedExampleTx
-    (mkDijkstraBasedExampleTxBody $ exampleConwayBasedTxBody exampleDijkstraCerts)
-    (DijkstraSpending $ AsIx 0)
-
--- | Reusable Tx builder for Dijkstra onwards with PlutusV4 script witness.
-mkDijkstraBasedExampleTx ::
+exampleDijkstraBasedTopTx ::
   forall era.
   ( AlonzoEraTx era
+  , DijkstraEraTxBody era
+  , Value era ~ MaryValue
+  , DijkstraEraScript era
+  , EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
+  , EraPlutusTxInfo PlutusV3 era
+  , EraPlutusTxInfo PlutusV4 era
   , AlonzoEraTxAuxData era
+  ) =>
+  Tx TopTx era
+exampleDijkstraBasedTopTx =
+  exampleConwayBasedTopTx
+    & addDijkstraBasedTxFeatures
+    & addDijkstraBasedTopTxFeatures
+
+exampleDijkstraBasedSubTx ::
+  forall era.
+  ( AlonzoEraTx era
+  , DijkstraEraTxBody era
+  , Value era ~ MaryValue
+  , DijkstraEraScript era
+  , EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
+  , EraPlutusTxInfo PlutusV3 era
+  , EraPlutusTxInfo PlutusV4 era
+  , AlonzoEraTxAuxData era
+  ) =>
+  Tx SubTx era
+exampleDijkstraBasedSubTx =
+  exampleConwayBasedTx
+    & addDijkstraBasedTxFeatures
+    & addDijkstraBasedSubTxFeatures
+
+addDijkstraBasedTopTxFeatures ::
+  forall era.
+  ( AlonzoEraTx era
+  , DijkstraEraTxBody era
+  , AlonzoEraTxAuxData era
+  , DijkstraEraScript era
   , EraPlutusTxInfo 'PlutusV1 era
   , EraPlutusTxInfo 'PlutusV2 era
   , EraPlutusTxInfo 'PlutusV3 era
   , EraPlutusTxInfo 'PlutusV4 era
+  , Value era ~ MaryValue
   ) =>
-  TxBody TopTx era ->
-  PlutusPurpose AsIx era ->
+  Tx TopTx era ->
   Tx TopTx era
-mkDijkstraBasedExampleTx txBody scriptPurpose =
-  mkConwayBasedExampleTx txBody scriptPurpose
+addDijkstraBasedTopTxFeatures tx =
+  tx
+    & bodyTxL . subTransactionsTxBodyL .~ OMap.fromFoldable [exampleDijkstraBasedSubTx]
+
+addDijkstraBasedSubTxFeatures ::
+  forall era.
+  ( AlonzoEraTx era
+  , DijkstraEraTxBody era
+  ) =>
+  Tx SubTx era ->
+  Tx SubTx era
+addDijkstraBasedSubTxFeatures tx =
+  tx
+    & bodyTxL . requiredTopLevelGuardsL
+      <>~ Map.fromList
+        [ (KeyHashObj $ mkKeyHash 212, SNothing)
+        , (ScriptHashObj $ mkScriptHash 213, SJust $ exampleDatum @era)
+        ]
+
+addDijkstraBasedTxFeatures ::
+  forall era l.
+  ( AlonzoEraTx era
+  , DijkstraEraTxBody era
+  , AlonzoEraTxAuxData era
+  , DijkstraEraScript era
+  , EraPlutusTxInfo 'PlutusV1 era
+  , EraPlutusTxInfo 'PlutusV4 era
+  , Value era ~ MaryValue
+  ) =>
+  Tx l era ->
+  Tx l era
+addDijkstraBasedTxFeatures tx =
+  tx
     & witsTxL
       <>~ ( mkBasicTxWits
-              & scriptTxWitsL
-                .~ Map.singleton
-                  (hashScript @era $ alwaysSucceeds @'PlutusV4 3)
-                  (alwaysSucceeds @'PlutusV4 3)
+              & scriptTxWitsL <>~ Map.fromElems hashScript [alwaysSucceeds @'PlutusV4 3]
+              & rdmrsTxWitsL <>~ redeemers
           )
-    & auxDataTxL
-      %~ fmap
-        ( \auxData ->
-            auxData
-              & plutusScriptsTxAuxDataL
-                <>~ Map.singleton PlutusV4 (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV4 3)))
-        )
-
-mkDijkstraBasedExampleTxBody ::
-  forall era.
-  ( DijkstraEraTxBody era
-  , AlonzoEraTxAuxData era
-  , EraTx era
-  , Value era ~ MaryValue
-  , EraPlutusTxInfo PlutusV4 era
-  , TxCert era ~ DijkstraTxCert era
-  ) =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkDijkstraBasedExampleTxBody txBody =
-  txBody
-    & outputsTxBodyL
+    & modifyTxAuxData
+      ( plutusScriptsTxAuxDataL
+          %~ Map.insertWith
+            (<>)
+            PlutusV4
+            (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV4 3)))
+      )
+    & bodyTxL . outputsTxBodyL
       <>~ StrictSeq.fromList
         [ mkBasicTxOut
             (mkAddr examplePayKey exampleStakeKey)
@@ -156,60 +202,21 @@ mkDijkstraBasedExampleTxBody txBody =
             & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
             & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV4 3)
         ]
-    & guardsTxBodyL
+    & bodyTxL . guardsTxBodyL
       .~ OSet.fromList
         [ KeyHashObj $ mkKeyHash 211
         , KeyHashObj $ mkKeyHash 212
         , ScriptHashObj $ mkScriptHash 213
         ]
-    & subTransactionsTxBodyL .~ OMap.fromFoldable [subTx]
-    & directDepositsTxBodyL .~ exampleDirectDeposits
-    & accountBalanceIntervalsTxBodyL .~ exampleAccountBalanceIntervals
+    & bodyTxL . directDepositsTxBodyL .~ exampleDirectDeposits
+    & bodyTxL . accountBalanceIntervalsTxBodyL .~ exampleAccountBalanceIntervals
   where
-    subTx =
-      mkBasicTx @era subTxBody
-        & witsTxL
-          <>~ ( mkBasicTxWits
-                  & scriptTxWitsL
-                    .~ Map.singleton
-                      (hashScript @era $ alwaysSucceeds @'PlutusV4 3)
-                      (alwaysSucceeds @'PlutusV4 3)
-              )
-        & auxDataTxL
-          %~ fmap
-            ( \auxData ->
-                auxData
-                  & plutusScriptsTxAuxDataL
-                    <>~ Map.singleton PlutusV4 (NE.singleton (plutusBinary (alwaysSucceedsPlutus @'PlutusV4 3)))
-            )
-    subTxBody =
-      mkBasicTxBody
-        & inputsTxBodyL .~ exampleTxIns
-        & referenceInputsTxBodyL .~ exampleTxIns
-        & outputsTxBodyL
-          <>~ StrictSeq.fromList
-            [ mkBasicTxOut
-                (mkAddr examplePayKey exampleStakeKey)
-                (exampleMultiAssetValue 2)
-                & datumTxOutL .~ Datum (dataToBinaryData exampleDatum)
-                & referenceScriptTxOutL .~ SJust (alwaysSucceeds @'PlutusV4 3)
-            ]
-        & vldtTxBodyL .~ ValidityInterval (SJust (SlotNo 2)) (SJust (SlotNo 4))
-        & mintTxBodyL .~ exampleMultiAsset 1
-        & withdrawalsTxBodyL .~ exampleWithdrawals
-        & auxDataHashTxBodyL .~ SJust (TxAuxDataHash $ mkDummySafeHash @EraIndependentTxAuxData 30)
-        & scriptIntegrityHashTxBodyL .~ SJust (mkDummySafeHash 42)
-        & networkIdTxBodyL .~ SJust Mainnet
-        & certsTxBodyL .~ exampleDijkstraCerts
-        & votingProceduresTxBodyL .~ exampleVotingProcedures
-        & proposalProceduresTxBodyL .~ OSet.singleton exampleProposalProcedure
-        & currentTreasuryValueTxBodyL .~ SJust (Coin 867530900000)
-        & treasuryDonationTxBodyL .~ Coin 1000000
-        & requiredTopLevelGuardsL
-          .~ Map.fromList
-            [ (KeyHashObj $ mkKeyHash 212, SNothing)
-            , (ScriptHashObj $ mkScriptHash 213, SJust $ exampleDatum @era)
-            ]
+    redeemers =
+      Redeemers $
+        Map.fromList
+          [ (GuardingPurpose $ AsIx 3, (redeemerData, ExUnits 5000 5000))
+          ]
+    redeemerData = Data @era (P.Constr 1 [P.List [P.I 1], P.Map [(P.I 2, P.B "2")]])
 
 exampleDirectDeposits :: DirectDeposits
 exampleDirectDeposits =
@@ -229,31 +236,3 @@ exampleAccountBalanceIntervals =
         , AccountBalanceBothBounds (Inclusive $ Coin 100) (Exclusive $ Coin 5000)
         )
       ]
-
-exampleDijkstraCerts :: StrictSeq.StrictSeq (DijkstraTxCert era)
-exampleDijkstraCerts =
-  StrictSeq.fromList
-    [ DijkstraTxCertDeleg $ DijkstraRegCert (KeyHashObj (mkKeyHash 1)) (Coin 2000000)
-    , DijkstraTxCertDeleg $ DijkstraUnRegCert (KeyHashObj (mkKeyHash 1)) (Coin 2000000)
-    , DijkstraTxCertDeleg $
-        DijkstraDelegCert (KeyHashObj (mkKeyHash 3)) (DelegStake (sppId exampleStakePoolParams))
-    , DijkstraTxCertDeleg $
-        DijkstraDelegCert (KeyHashObj (mkKeyHash 4)) (DelegVote (DRepKeyHash (mkKeyHash 5)))
-    , DijkstraTxCertDeleg $
-        DijkstraDelegCert
-          (KeyHashObj (mkKeyHash 6))
-          (DelegStakeVote (sppId exampleStakePoolParams) (DRepKeyHash (mkKeyHash 7)))
-    , DijkstraTxCertDeleg $
-        DijkstraRegDelegCert
-          (KeyHashObj (mkKeyHash 8))
-          (DelegStake (sppId exampleStakePoolParams))
-          (Coin 2000000)
-    , DijkstraTxCertPool $ RegPool exampleStakePoolParams
-    , DijkstraTxCertPool $ RetirePool (sppId exampleStakePoolParams) (EpochNo 2)
-    , DijkstraTxCertGov $ ConwayRegDRep (KeyHashObj (mkKeyHash 9)) (Coin 500000) (SJust exampleAnchor)
-    , DijkstraTxCertGov $ ConwayUnRegDRep (KeyHashObj (mkKeyHash 9)) (Coin 500000)
-    , DijkstraTxCertGov $ ConwayUpdateDRep (KeyHashObj (mkKeyHash 9)) (SJust exampleAnchor)
-    , DijkstraTxCertGov $
-        ConwayAuthCommitteeHotKey (KeyHashObj (mkKeyHash 10)) (KeyHashObj (mkKeyHash 11))
-    , DijkstraTxCertGov $ ConwayResignCommitteeColdKey (KeyHashObj (mkKeyHash 10)) (SJust exampleAnchor)
-    ]