+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/Examples.hs
@@ -6,14 +6,16 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
+-- | The example transactions in this module are not valid transactions. We
+-- don't care, we are only interested in serialisation, not validation.
 module Test.Cardano.Ledger.Mary.Examples (
   ledgerExamples,
-  exampleMaryBasedShelleyTxBody,
-  exampleMaryBasedTxBody,
+  exampleMaryBasedTx,
   exampleMultiAsset,
   exampleMultiAssetValue,
 ) where
 
+import Cardano.Ledger.Allegra.Scripts (AllegraEraScript)
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Genesis (NoGenesis (..))
 import Cardano.Ledger.Mary (ApplyTxError (MaryApplyTxError), MaryEra)
@@ -26,14 +28,21 @@ import Cardano.Ledger.Shelley.Rules (
   ShelleyLedgerPredFailure (DelegsFailure),
  )
 import qualified Data.Map.Strict as Map (singleton)
+import qualified Data.Sequence.Strict as StrictSeq
+import Data.Typeable (Typeable)
 import Lens.Micro
 import Test.Cardano.Ledger.Allegra.Examples (
-  exampleAllegraBasedShelleyTxBody,
-  exampleAllegraBasedTxBody,
-  mkAllegraBasedExampleTx,
+  exampleAllegraBasedTx,
  )
+import Test.Cardano.Ledger.Core.KeyPair (mkAddr)
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples,
+  addShelleyBasedTopTxExampleFee,
+  addShelleyToBabbageExampleProposedPUpdates,
+  addShelleyToBabbageTxCerts,
+  addShelleyToConwayTxCerts,
+  examplePayKey,
+  exampleStakeKey,
   mkKeyHash,
   mkScriptHash,
   mkShelleyBasedLedgerExamples,
@@ -46,36 +55,37 @@ ledgerExamples =
         DelegateeNotRegisteredDELEG @MaryEra (mkKeyHash 1)
     )
     (exampleMultiAssetValue 1)
-    (mkAllegraBasedExampleTx exampleMaryBasedShelleyTxBody)
+    exampleMaryTx
     NoGenesis
+  where
+    exampleMaryTx :: Tx TopTx MaryEra
+    exampleMaryTx =
+      exampleMaryBasedTx
+        & addShelleyBasedTopTxExampleFee
+        & addShelleyToBabbageExampleProposedPUpdates
+        & addShelleyToBabbageTxCerts
+        & addShelleyToConwayTxCerts
 
-exampleMaryBasedShelleyTxBody ::
-  forall era.
-  ( MaryEraTxBody era
-  , ShelleyEraTxBody era
+-- Complete transaction which is compatible with any era starting with Mary.
+-- This transaction forms the basis on which future era transactions will be
+-- at the very least based on.
+exampleMaryBasedTx ::
+  forall era l.
+  ( EraTx era
+  , MaryEraTxBody era
   , Value era ~ MaryValue
+  , AllegraEraTxAuxData era
+  , AllegraEraScript era
+  , Typeable l
   ) =>
-  TxBody TopTx era
-exampleMaryBasedShelleyTxBody =
-  mkMaryBasedExampleTxBody (exampleAllegraBasedShelleyTxBody (exampleMultiAssetValue 1))
-
-exampleMaryBasedTxBody ::
-  forall era.
-  ( MaryEraTxBody era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era
-exampleMaryBasedTxBody =
-  mkMaryBasedExampleTxBody (exampleAllegraBasedTxBody (exampleMultiAssetValue 1))
-
-mkMaryBasedExampleTxBody ::
-  forall era.
-  MaryEraTxBody era =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkMaryBasedExampleTxBody txBody =
-  txBody
-    & mintTxBodyL .~ exampleMultiAsset 1
+  Tx l era
+exampleMaryBasedTx =
+  exampleAllegraBasedTx
+    & bodyTxL . outputsTxBodyL
+      <>~ StrictSeq.fromList
+        [ mkBasicTxOut (mkAddr examplePayKey exampleStakeKey) $ exampleMultiAssetValue 1
+        ]
+    & bodyTxL . mintTxBodyL .~ exampleMultiAsset 1
 
 exampleMultiAssetValue :: Int -> MaryValue
 exampleMultiAssetValue x = MaryValue (Coin 100) $ exampleMultiAsset x