+++ eras/mary/impl/CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## 1.10.1.0
 
+* Add `ApplyTick` instance for `MaryEra`
+* Add `EraForecast` and `ShelleyEraForecast` instances for `MaryEra`.
+
 ### `cddl`
 
 * Change `ipv4` and `ipv6` to use exact byte sizes (4 and 16 respectively), no longer allowing leftover bytes
-* Add `ApplyTick` instance for `MaryEra`
-* Add `EraForecast` and `ShelleyEraForecast` instances for `MaryEra`.
+
+### `testlib`
+
+* In `Test.Cardano.Ledger.Mary.Examples`:
+  - Remove `exampleMaryBasedShelleyTxBody`, `exampleMaryBasedTxBody`
+  - Add `exampleMaryBasedTx`
 
 ## 1.10.0.0